Sports class separated into multiple classes and subclasses
