+++ DataDownload.py
@@ -1,10 +1,11 @@
 from Tools.Sports import sportclass, sports
+from Tools.FileManager.manager import FileManager
 
 def download_data():
 
     for value in sports.sports_dic:
         sport = sportclass.Sport(value, sports.sports_dic[value])
-        sport.set_up_sport()
+        fm = FileManager(sport.get_sport(), sport.get_version())
+        fm.set_up_sport()
 
 download_data()
-
